requirements.txt
streamlit
pandas
prophet
plotly
openai
python-dotenv